feat: create new standup with custom list of people
go to /new to create the list of people. When you press the button to start the standup, it will redirect you to a route with the standup. The create route can be sent to anyone to let them join the standup.
