+++ src/components/standup-component/standup-component.css
