+++ src/root.tsx
@@ -7,7 +7,7 @@ import {
 import { RouterHead } from "./components/router-head/router-head";
 
 import globalStyles from "./global.css?inline";
-import { standupStatesContext } from "./routes/layout";
+import { standupParticipantsContext } from "./shared/standup-participants.context";
 
 export default component$(() => {
   /**
@@ -17,7 +17,7 @@ export default component$(() => {
    * Dont remove the `<head>` and `<body>` elements.
    */
   useStyles$(globalStyles);
-  useContextProvider(standupStatesContext, {});
+  useContextProvider(standupParticipantsContext, {});
 
   return (
     <QwikCityProvider>
