+++ src/routes/[standupId]/index.tsx
@@ -1,23 +1,17 @@
-import {
-  component$,
-  $,
-  useStylesScoped$,
-  useContext,
-  useStore,
-} from "@builder.io/qwik";
+import { component$, useContext, useStore } from "@builder.io/qwik";
 import { DocumentHead, useLocation } from "@builder.io/qwik-city";
-import { PersonStatus } from "~/components/person-status/person-status";
+import { StandupComponent } from "~/components/standup-component/standup-component";
 import { useSyncedStandupState } from "~/hooks/useSyncedStandupState";
-import { Person, PersonState, StandupState } from "~/shared/types";
-import styles from "./index.css?inline";
-import { standupStatesContext } from "../layout";
+import { Person, StandupState } from "~/shared/standup-state.types";
+import { standupParticipantsContext } from "../../shared/standup-participants.context";
 
 export const makeInitialStandupState = (
   people: Person[],
   id: string
 ): StandupState => {
   return {
-    ...initialStandupState,
+    orderPosition: 0,
+    allDone: false,
     standupId: id,
     people: people.map((pers, index) => ({
       done: false,
@@ -28,113 +22,32 @@ export const makeInitialStandupState = (
   };
 };
 
-export const makeInitialPeopleState = (names: string[]): PersonState[] =>
-  names.map((name, index) => ({ done: false, name, order: index, id: index }));
-
-const initialPeople = makeInitialPeopleState([
-  "Ben",
-  "Chaz",
-  "Jason",
-  "Jerry",
-  "Joe",
-  "Juana",
-  "Luke",
-  "Zack",
-]);
-
-export const initialStandupState: StandupState = {
-  orderPosition: 0,
-  people: initialPeople,
-  allDone: false,
-  standupId: "literalpie-open-standup",
+export const isStandupState = (
+  standup: StandupState | { standupId: StandupState["standupId"] }
+): standup is StandupState => {
+  return (standup as StandupState).people !== undefined;
 };
 
 export default component$(() => {
   const location = useLocation();
-  const standupStates = useContext(standupStatesContext);
+  const standupStates = useContext(standupParticipantsContext);
   const peopleForStandupId = standupStates[location.params.standupId];
-  console.log(
-    "people for this ID",
-    standupStates,
-    location.params.standupId,
-    peopleForStandupId
-  );
-  const thisStandup: Partial<StandupState> = peopleForStandupId
+  const matchingStandup:
+    | StandupState
+    | { standupId: StandupState["standupId"] } = peopleForStandupId
     ? makeInitialStandupState(peopleForStandupId, location.params.standupId)
     : { standupId: location.params.standupId };
-  const standupState = useStore(thisStandup, { recursive: true });
-  useStylesScoped$(styles);
+  const standupState = useStore(matchingStandup, { recursive: true });
   useSyncedStandupState(standupState);
-  // calculates on every render, but okay because people will never be large
-  const currentPerson = standupState.people?.find(
-    (person) => person.order === standupState.orderPosition
-  );
-
-  const setNextOrderPosition = $(() => {
-    if (standupState.allDone) {
-      return;
-    }
-    const sortedPeople =
-      standupState.people?.sort((a, b) => a.order - b.order) ?? [];
-    const nextAfterCurrent = sortedPeople.find(
-      (person, index) =>
-        standupState.orderPosition !== undefined &&
-        index > standupState.orderPosition &&
-        !person.done
-    );
-    const firstNotDone = sortedPeople.find((person) => !person.done);
-    const nextOrderPosition = nextAfterCurrent?.order ?? firstNotDone?.order;
-    standupState.orderPosition = nextOrderPosition;
-    standupState.allDone = nextOrderPosition === undefined;
-  });
 
   return (
-    <div>
-      {standupState.people?.map((person) => {
-        return (
-          <PersonStatus
-            key={person.name}
-            name={person.name}
-            done={person.done}
-            current={person.order === standupState.orderPosition}
-          />
-        );
-      })}
-      <div class="button-container">
-        {standupState.allDone ? (
-          <>
-            <div class="flex-grow">All Done!</div>
-            <button
-              class="flex-grow"
-              onClick$={() => {
-                standupState.allDone = initialStandupState.allDone;
-                standupState.orderPosition = initialStandupState.orderPosition;
-                standupState.people?.forEach((person) => (person.done = false));
-              }}
-            >
-              Reset
-            </button>
-          </>
-        ) : (
-          <>
-            <button
-              class="flex-grow"
-              onClick$={() => {
-                if (currentPerson) {
-                  currentPerson.done = true;
-                }
-                setNextOrderPosition();
-              }}
-            >
-              Next
-            </button>
-            <button class="flex-grow" onClick$={setNextOrderPosition}>
-              Skip
-            </button>
-          </>
-        )}
-      </div>
-    </div>
+    <>
+      {isStandupState(standupState) ? (
+        <StandupComponent standupState={standupState} />
+      ) : (
+        <div>Standup Not Found</div>
+      )}
+    </>
   );
 });
