+++ src/routes/layout.tsx
@@ -1,9 +1,5 @@
-import { component$, Slot, createContext } from "@builder.io/qwik";
-import { Person } from "~/shared/types";
+import { component$, Slot } from "@builder.io/qwik";
 import Header from "../components/header/header";
-
-export const standupStatesContext =
-  createContext<Record<string, Person[]>>("standup-state");
 
 export default component$(() => {
   return (
