+++ src/routes/new/index.tsx
@@ -7,8 +7,8 @@ import {
   useContext,
 } from "@builder.io/qwik";
 import { useNavigate } from "@builder.io/qwik-city";
-import { Person } from "~/shared/types";
-import { standupStatesContext } from "../layout";
+import { Person } from "~/shared/standup-state.types";
+import { standupParticipantsContext } from "../../shared/standup-participants.context";
 import styles from "./index.css?inline";
 
 export default component$(() => {
@@ -16,7 +16,7 @@ export default component$(() => {
   const participants = useSignal<Person[]>([]);
   const newPartic = useSignal<string>();
   const navigate = useNavigate();
-  const standupState = useContext(standupStatesContext);
+  const standupState = useContext(standupParticipantsContext);
   const submitNewParticipant = $(() => {
     if (!newPartic.value) {
       return;
